clean up routing and add redirects for state index pages
