+++ components/EventPages/ComingSoon.js
@@ -1,0 +1,17 @@
+/**
+ * Page to render coming soon for an event we want to tease
+ */
+
+import React from 'react';
+import styled from 'styled-components';
+
+const ComingSoon = () => {
+  return (
+    <p>
+      This is a coming soon message for an event we want to tease but don't have
+      all the details for yet
+    </p>
+  );
+};
+
+export default styled(ComingSoon)``;
